+++ CHANGES.md
@@ -6,6 +6,11 @@ Version 0.3.0
 
 To be released.
 
+ -  Added `--no-word-wrap` to `sacho preview` and `sacho show` for publishing
+    release notes without Sacho's canonical 80-column wrapping.  [[#3]]
+
+[#3]: https://github.com/dahlia/sacho/pull/3
+
 
 Version 0.2.0
 -------------